requirements/base.txt
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
Markdown==2.6.11
psycopg2==2.7.4
pytz==2018.3
